Cleanup ResolvedConfigurationIdentifier
Rename ResolvedConfigurationIdentifier to ComponentVariantIdentifier
Avoid almost all usages of ComponentVariantIdentifier, since it is identified by variant name and we should avoid using the variant name in the engine
Cleanup NodeState constructor to avoid mutating passed-in component
Remove nullable parameters from NodeState constructor which were only null in tests
Add javadoc to clarify we should avoid using the variant's name where possible
