+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/DefaultResolvedDependency.java
@@ -21,6 +21,7 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ListMultimap;
 import org.apache.commons.lang.ObjectUtils;
 import org.gradle.api.InvalidUserDataException;
+import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.ResolvedArtifact;
 import org.gradle.api.artifacts.ResolvedDependency;
 import org.gradle.api.artifacts.ResolvedModuleVersion;
@@ -37,6 +38,7 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.LinkedHashSet;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -44,16 +46,16 @@ public class DefaultResolvedDependency implements ResolvedDependency, Dependency
     private final Set<DefaultResolvedDependency> children = new LinkedHashSet<>();
     private final Set<ResolvedDependency> parents = new LinkedHashSet<>();
     private final ListMultimap<ResolvedDependency, ResolvedArtifactSet> parentArtifacts = ArrayListMultimap.create();
-    private final String name;
-    private final ResolvedConfigurationIdentifier resolvedConfigId;
+    private final String variantName;
+    private final ModuleVersionIdentifier moduleVersionId;
     private final BuildOperationExecutor buildOperationProcessor;
     private final Set<ResolvedArtifactSet> moduleArtifacts;
     private final Map<ResolvedDependency, Set<ResolvedArtifact>> allArtifactsCache = new HashMap<>();
     private Set<ResolvedArtifact> allModuleArtifactsCache;
 
-    public DefaultResolvedDependency(ResolvedConfigurationIdentifier resolvedConfigurationIdentifier, BuildOperationExecutor buildOperationProcessor) {
-        this.name = String.format("%s:%s:%s", resolvedConfigurationIdentifier.getModuleGroup(), resolvedConfigurationIdentifier.getModuleName(), resolvedConfigurationIdentifier.getModuleVersion());
-        this.resolvedConfigId = resolvedConfigurationIdentifier;
+    public DefaultResolvedDependency(String variantName, ModuleVersionIdentifier moduleVersionId, BuildOperationExecutor buildOperationProcessor) {
+        this.moduleVersionId = moduleVersionId;
+        this.variantName = variantName;
         this.buildOperationProcessor = buildOperationProcessor;
         this.moduleArtifacts = new LinkedHashSet<>();
     }
@@ -65,32 +67,32 @@ public class DefaultResolvedDependency implements ResolvedDependency, Dependency
 
     @Override
     public String getName() {
-        return name;
+        return String.format("%s:%s:%s", moduleVersionId.getGroup(), moduleVersionId.getName(), moduleVersionId.getVersion());
     }
 
     @Override
     public String getModuleGroup() {
-        return resolvedConfigId.getModuleGroup();
+        return moduleVersionId.getGroup();
     }
 
     @Override
     public String getModuleName() {
-        return resolvedConfigId.getModuleName();
+        return moduleVersionId.getName();
     }
 
     @Override
     public String getModuleVersion() {
-        return resolvedConfigId.getModuleVersion();
+        return moduleVersionId.getVersion();
     }
 
     @Override
     public String getConfiguration() {
-        return resolvedConfigId.getConfiguration();
+        return variantName;
     }
 
     @Override
     public ResolvedModuleVersion getModule() {
-        return new DefaultResolvedModuleVersion(resolvedConfigId.getId());
+        return new DefaultResolvedModuleVersion(moduleVersionId);
     }
 
     @Override
@@ -172,7 +174,7 @@ public class DefaultResolvedDependency implements ResolvedDependency, Dependency
 
     @Override
     public String toString() {
-        return name + ";" + getConfiguration();
+        return getName() + ";" + getConfiguration();
     }
 
     @Override
@@ -185,12 +187,13 @@ public class DefaultResolvedDependency implements ResolvedDependency, Dependency
         }
 
         DefaultResolvedDependency that = (DefaultResolvedDependency) o;
-        return resolvedConfigId.equals(that.resolvedConfigId);
+        return Objects.equals(variantName, that.variantName) &&
+            Objects.equals(moduleVersionId, that.moduleVersionId);
     }
 
     @Override
     public int hashCode() {
-        return resolvedConfigId.hashCode();
+        return variantName.hashCode() ^ moduleVersionId.hashCode();
     }
 
     public void addChild(DefaultResolvedDependency child) {
