+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/DependencyGraphNode.java
@@ -16,7 +16,6 @@
 
 package org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph;
 
-import org.gradle.api.internal.artifacts.ResolvedConfigurationIdentifier;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.ComponentResolutionState;
 import org.gradle.internal.component.local.model.LocalFileDependencyMetadata;
 import org.gradle.internal.component.model.VariantGraphResolveMetadata;
@@ -29,8 +28,6 @@ import java.util.Set;
  */
 public interface DependencyGraphNode extends ResolvedGraphVariant {
     boolean isRoot();
-
-    ResolvedConfigurationIdentifier getResolvedConfigurationId();
 
     DependencyGraphComponent getOwner();
