+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/conflicts/DefaultCapabilitiesConflictHandler.java
@@ -209,7 +209,7 @@ public class DefaultCapabilitiesConflictHandler implements CapabilitiesConflictH
 
                             @Override
                             public String getVariantName() {
-                                return node.getResolvedConfigurationId().getConfiguration();
+                                return node.getMetadata().getName();
                             }
 
                             @Override
