+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/oldresult/TransientConfigurationResultsBuilder.java
@@ -18,11 +18,11 @@
 package org.gradle.api.internal.artifacts.ivyservice.resolveengine.oldresult;
 
 import org.gradle.api.artifacts.Dependency;
+import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.internal.artifacts.DefaultResolvedDependency;
 import org.gradle.api.internal.artifacts.DependencyGraphNodeResult;
 import org.gradle.api.internal.artifacts.ImmutableModuleIdentifierFactory;
-import org.gradle.api.internal.artifacts.ResolvedConfigurationIdentifier;
-import org.gradle.api.internal.artifacts.ResolvedConfigurationIdentifierSerializer;
+import org.gradle.api.internal.artifacts.ModuleVersionIdentifierSerializer;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvedArtifactSet;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.SelectedArtifactResults;
 import org.gradle.api.logging.Logger;
@@ -59,21 +59,22 @@ public class TransientConfigurationResultsBuilder {
     private final BinaryStore binaryStore;
     private final Store<TransientConfigurationResults> cache;
     private final BuildOperationExecutor buildOperationProcessor;
-    private final ResolvedConfigurationIdentifierSerializer resolvedConfigurationIdentifierSerializer;
+    private final ModuleVersionIdentifierSerializer moduleVersionIdSerializer;
     private BinaryStore.BinaryData binaryData;
 
     public TransientConfigurationResultsBuilder(BinaryStore binaryStore, Store<TransientConfigurationResults> cache, ImmutableModuleIdentifierFactory moduleIdentifierFactory, BuildOperationExecutor buildOperationProcessor) {
-        this.resolvedConfigurationIdentifierSerializer = new ResolvedConfigurationIdentifierSerializer(moduleIdentifierFactory);
+        this.moduleVersionIdSerializer = new ModuleVersionIdentifierSerializer(moduleIdentifierFactory);
         this.binaryStore = binaryStore;
         this.cache = cache;
         this.buildOperationProcessor = buildOperationProcessor;
     }
 
-    public void resolvedDependency(final Long id, final ResolvedConfigurationIdentifier details) {
+    public void resolvedDependency(final Long id, ModuleVersionIdentifier moduleVersionId, String variantName) {
         binaryStore.write(encoder -> {
             encoder.writeByte(NODE);
             encoder.writeSmallLong(id);
-            resolvedConfigurationIdentifierSerializer.write(encoder, details);
+            moduleVersionIdSerializer.write(encoder, moduleVersionId);
+            encoder.writeString(variantName);
         });
     }
 
@@ -142,8 +143,9 @@ public class TransientConfigurationResultsBuilder {
                 switch (type) {
                     case NODE:
                         id = decoder.readSmallLong();
-                        ResolvedConfigurationIdentifier details = resolvedConfigurationIdentifierSerializer.read(decoder);
-                        allDependencies.put(id, new DefaultResolvedDependency(details, buildOperationProcessor));
+                        ModuleVersionIdentifier moduleVersionId = moduleVersionIdSerializer.read(decoder);
+                        String variantName = decoder.readString();
+                        allDependencies.put(id, new DefaultResolvedDependency(variantName, moduleVersionId, buildOperationProcessor));
                         break;
                     case ROOT:
                         id = decoder.readSmallLong();
