+++ platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/DefaultResolvedDependencySpec.groovy
@@ -15,15 +15,178 @@
  */
 package org.gradle.api.internal.artifacts
 
+import org.gradle.api.InvalidUserDataException
+import org.gradle.api.artifacts.ModuleVersionIdentifier
 import org.gradle.api.artifacts.ResolvedArtifact
+import org.gradle.api.artifacts.component.ComponentArtifactIdentifier
 import org.gradle.api.internal.attributes.ImmutableAttributes
+import org.gradle.api.internal.tasks.TaskDependencyContainer
+import org.gradle.internal.Describables
+import org.gradle.internal.component.model.IvyArtifactName
+import org.gradle.internal.operations.BuildOperationExecutor
 import org.gradle.internal.operations.TestBuildOperationExecutor
+import org.gradle.util.TestUtil
 import spock.lang.Specification
 
-class DefaultResolvedDependencySpec extends Specification {
-    final dependency = new DefaultResolvedDependency(new ResolvedConfigurationIdentifier(DefaultModuleVersionIdentifier.newId("group", "module", "version"), "config"), new TestBuildOperationExecutor())
+import static com.google.common.collect.Iterables.concat
+import static com.google.common.collect.Sets.newHashSet
+import static org.gradle.util.Matchers.strictlyEquals
+import static org.gradle.util.internal.WrapUtil.toSet
 
+class DefaultResolvedDependencySpec extends Specification {private BuildOperationExecutor buildOperationProcessor = new TestBuildOperationExecutor()
+
+    def init() {
+        when:
+        String someGroup = "someGroup"
+        String someName = "someName"
+        String someVersion = "someVersion"
+        String someConfiguration = "someConfiguration"
+        DefaultResolvedDependency resolvedDependency = new DefaultResolvedDependency(someConfiguration, newId(someGroup, someName, someVersion), buildOperationProcessor)
+
+        then:
+        resolvedDependency.name == someGroup + ":" + someName + ":" + someVersion
+        resolvedDependency.moduleGroup == someGroup
+        resolvedDependency.moduleName == someName
+        resolvedDependency.moduleVersion == someVersion
+        resolvedDependency.configuration == someConfiguration
+        resolvedDependency.moduleArtifacts.empty
+        resolvedDependency.children.empty
+        resolvedDependency.parents.empty
+    }
+
+    def getAllModuleArtifactsReturnsUnionOfAllIncomingArtifacts() {
+        ResolvedArtifact artifact1 = createArtifact("1")
+        ResolvedArtifact artifact2 = createArtifact("2")
+
+        when:
+        DefaultResolvedDependency resolvedDependency = new DefaultResolvedDependency("someConfiguration", newId("someGroup", "someName", "someVersion"), buildOperationProcessor)
+
+        DefaultResolvedDependency parent1 = new DefaultResolvedDependency("p1", newId("someGroup", "someChild", "someVersion"), buildOperationProcessor)
+        parent1.addChild(resolvedDependency)
+
+        DefaultResolvedDependency parent2 = new DefaultResolvedDependency("p2", newId("someGroup", "someChild", "someVersion"), buildOperationProcessor)
+        parent2.addChild(resolvedDependency)
+
+        resolvedDependency.addParentSpecificArtifacts(parent1, TestArtifactSet.create(ImmutableAttributes.EMPTY, Collections.singleton(artifact2)))
+        resolvedDependency.addParentSpecificArtifacts(parent2, TestArtifactSet.create(ImmutableAttributes.EMPTY, Arrays.asList(artifact1, artifact2)))
+
+        then:
+        resolvedDependency.allModuleArtifacts == [artifact1, artifact2] as Set
+    }
+
+    def getParentArtifacts() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+
+        Set<ResolvedArtifact> parent1SpecificArtifacts = toSet(createArtifact("parent1Specific"))
+        DefaultResolvedDependency parentResolvedDependency1 = createAndAddParent("parent1", resolvedDependency, parent1SpecificArtifacts)
+
+        Set<ResolvedArtifact> parent2SpecificArtifacts = toSet(createArtifact("parent2Specific"))
+        DefaultResolvedDependency parentResolvedDependency2 = createAndAddParent("parent2", resolvedDependency, parent2SpecificArtifacts)
+
+        then:
+        resolvedDependency.getParentArtifacts(parentResolvedDependency1) == parent1SpecificArtifacts
+        resolvedDependency.getParentArtifacts(parentResolvedDependency2) == parent2SpecificArtifacts
+    }
+
+    def getArtifacts() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+
+        def parent1SpecificArtifacts = toSet(createArtifact("parent1Specific"))
+        def parentResolvedDependency1 = createAndAddParent("parent1", resolvedDependency, parent1SpecificArtifacts)
+
+        then:
+        resolvedDependency.getArtifacts(parentResolvedDependency1) == parent1SpecificArtifacts
+    }
+
+    def getArtifactsWithParentWithoutParentArtifacts() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+
+        DefaultResolvedDependency parent = new DefaultResolvedDependency("someConfiguration", newId("someGroup", "parent", "someVersion"), buildOperationProcessor)
+        resolvedDependency.getParents().add(parent)
+
+        then:
+        resolvedDependency.getArtifacts(parent).empty
+    }
+
+    def getParentArtifactsWithParentWithoutParentArtifacts() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+
+        DefaultResolvedDependency parent = new DefaultResolvedDependency("someConfiguration", newId("someGroup", "parent", "someVersion"), buildOperationProcessor)
+        resolvedDependency.getParents().add(parent)
+
+        then:
+        resolvedDependency.getParentArtifacts(parent).empty
+    }
+
+    def getParentArtifactsWithUnknownParent() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+        DefaultResolvedDependency unknownParent = new DefaultResolvedDependency("someConfiguration", newId("someGroup", "parent2", "someVersion"), buildOperationProcessor)
+
+        resolvedDependency.getParentArtifacts(unknownParent)
+
+        then:
+        thrown(InvalidUserDataException)
+    }
+
+    def getArtifactsWithUnknownParent() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency();
+        DefaultResolvedDependency unknownParent = new DefaultResolvedDependency("someConfiguration", newId("someGroup", "parent2", "someVersion"), buildOperationProcessor);
+
+        resolvedDependency.getAllArtifacts(unknownParent)
+
+        then:
+        thrown(InvalidUserDataException)
+    }
+
+    def getAllArtifacts() {
+        when:
+        DefaultResolvedDependency resolvedDependency = createResolvedDependency()
+
+        Set<ResolvedArtifact> parent1SpecificArtifacts = newHashSet(createArtifact("parent1Specific"))
+        DefaultResolvedDependency parentResolvedDependency1 = createAndAddParent("parent1", resolvedDependency, parent1SpecificArtifacts)
+
+        createAndAddParent("parent2", resolvedDependency, newHashSet(createArtifact("parent2Specific")))
+
+        DefaultResolvedDependency child = new DefaultResolvedDependency("someChildConfiguration", newId("someGroup", "someChild", "someVersion"), buildOperationProcessor)
+        resolvedDependency.addChild(child)
+
+        Set<ResolvedArtifact> childParent1SpecificArtifacts = newHashSet(createArtifact("childParent1Specific"))
+        createAndAddParent("childParent1", child, childParent1SpecificArtifacts)
+
+        Set<ResolvedArtifact> childParent2SpecificArtifacts = newHashSet(createArtifact("childParent2Specific"))
+        createAndAddParent("childParent2", child, childParent2SpecificArtifacts)
+
+        then:
+        def expectedArtifacts = concat(parent1SpecificArtifacts, childParent1SpecificArtifacts, childParent2SpecificArtifacts) as Set
+        resolvedDependency.getAllArtifacts(parentResolvedDependency1) == expectedArtifacts
+    }
+
+    def equalsAndHashCode() {
+        when:
+        DefaultResolvedDependency dependency = new DefaultResolvedDependency("config", newId("group", "name", "version"), buildOperationProcessor)
+        DefaultResolvedDependency same = new DefaultResolvedDependency("config", newId("group", "name", "version"), buildOperationProcessor)
+        DefaultResolvedDependency differentGroup = new DefaultResolvedDependency("config", newId("other", "name", "version"), buildOperationProcessor)
+        DefaultResolvedDependency differentName = new DefaultResolvedDependency("config", newId("group", "other", "version"), buildOperationProcessor)
+        DefaultResolvedDependency differentVersion = new DefaultResolvedDependency("config", newId("group", "name", "other"), buildOperationProcessor)
+        DefaultResolvedDependency differentConfiguration = new DefaultResolvedDependency("other", newId("group", "name", "version"), buildOperationProcessor)
+
+        then:
+        strictlyEquals(dependency, same)
+        dependency != differentGroup
+        dependency != differentName
+        dependency != differentVersion
+        dependency != differentConfiguration
+    }
     def "provides meta-data about the module"() {
+        given:
+        def dependency = new DefaultResolvedDependency("config", newId("group", "module", "version"), new TestBuildOperationExecutor())
+
         expect:
         dependency.module.id.group == "group"
         dependency.module.id.name == "module"
@@ -40,6 +203,8 @@ class DefaultResolvedDependencySpec extends Specification {
         ResolvedArtifact artifact7 = artifact("c", "a-classifier", "jar", "jar")
 
         given:
+        def dependency = new DefaultResolvedDependency("config", newId("group", "module", "version"), new TestBuildOperationExecutor())
+
         add(dependency, artifact6)
         add(dependency, artifact1)
         add(dependency, artifact3)
@@ -57,6 +222,8 @@ class DefaultResolvedDependencySpec extends Specification {
         ResolvedArtifact artifact2 = artifact("a", null, "jar", "jar")
 
         given:
+        def dependency = new DefaultResolvedDependency("config", newId("group", "module", "version"), new TestBuildOperationExecutor())
+
         add(dependency, artifact1)
         add(dependency, artifact2)
 
@@ -75,6 +242,8 @@ class DefaultResolvedDependencySpec extends Specification {
         DefaultResolvedDependency parent = Mock()
 
         given:
+        def dependency = new DefaultResolvedDependency("config", newId("group", "module", "version"), new TestBuildOperationExecutor())
+
         dependency.parents.add(parent)
         dependency.addParentSpecificArtifacts(parent, TestArtifactSet.create(ImmutableAttributes.EMPTY, [artifact6, artifact1, artifact7, artifact5, artifact2, artifact3, artifact4]))
 
@@ -94,5 +263,33 @@ class DefaultResolvedDependencySpec extends Specification {
 
     def add(DefaultResolvedDependency dependency, ResolvedArtifact artifact) {
         dependency.addParentSpecificArtifacts(Stub(DefaultResolvedDependency), TestArtifactSet.create(ImmutableAttributes.EMPTY, [artifact]))
+    }
+
+    private DefaultResolvedDependency createResolvedDependency() {
+        return new DefaultResolvedDependency("someConfiguration", newId("someGroup", "someName", "someVersion"), buildOperationProcessor)
+    }
+
+    private DefaultResolvedDependency createAndAddParent(String parentName, DefaultResolvedDependency resolvedDependency, Set<ResolvedArtifact> parentSpecificArtifacts) {
+        DefaultResolvedDependency parent = new DefaultResolvedDependency("someConfiguration", newId("someGroup", parentName, "someVersion"), buildOperationProcessor)
+        resolvedDependency.getParents().add(parent)
+        resolvedDependency.addParentSpecificArtifacts(parent, TestArtifactSet.create(ImmutableAttributes.EMPTY, parentSpecificArtifacts))
+        return parent
+    }
+
+    static ModuleVersionIdentifier newId(String group, String name, String version) {
+        return DefaultModuleVersionIdentifier.newId(group, name, version)
+    }
+
+    private ResolvedArtifact createArtifact(String name) {
+        def id = DefaultModuleVersionIdentifier.newId("group", name, "1.2")
+        IvyArtifactName artifactStub = Mock() {
+            getName() >> name
+            getType() >> "someType"
+            getExtension() >> "someExt"
+            getClassifier() >> null
+        }
+        def calculatedValueContainerFactory = TestUtil.calculatedValueContainerFactory()
+        def artifactSource = calculatedValueContainerFactory.create(Describables.of("artifact"), new File("pathTo" + name))
+        return new DefaultResolvableArtifact(id, artifactStub, Mock(ComponentArtifactIdentifier), Mock(TaskDependencyContainer), artifactSource, calculatedValueContainerFactory).toPublicView()
     }
 }
