+++ platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/builder/NodeStateTest.groovy
@@ -224,7 +224,7 @@ class NodeStateTest extends Specification {
         state.metadata >> metadata
         def resolveState = Stub(ResolveState)
 
-        def newState = new NodeState(idIdx++, null, Stub(ComponentState), resolveState, state, true)
+        def newState = new NodeState(idIdx++, Stub(ComponentState), resolveState, state, true)
         // if there are outgoing endorsing edges, also include a normal edge to make sure that it is filtered out
         metadata.dependencies >> ((0..<outgoingEndorsing).collect { edge(newState).dependencyMetadata } + (outgoingEndorsing > 0 ? [edge(newState, false).dependencyMetadata] : []))
         resolveState.moduleExclusions >> Mock(ModuleExclusions)
